Routes, Links, refactoring
