+++ frontend/src/App.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component, Fragment } from 'react';
 import './App.css';
 import {
   Switch,
@@ -6,14 +6,26 @@ import {
   Link,
   Redirect
 } from "react-router-dom";
-
-
+import HomePage from './pages/Homepage/Homepage';
+import MainPage from './pages/MainPage/MainPage';
+import Signup from './pages/Auth/Signup/Signup';
+import Login from './pages/Auth/Login/Login';
+import Profile from './pages/Profile/Profile';
+import Challenge from './pages/Challenge/Challenge';
 
 export class App extends Component {
   render() {
     return (
-      <div>
-      </div>
+      <Fragment>
+        <Switch>
+          <Route exact path="/" component={HomePage} />
+          <Route exact path="/main" component={MainPage} />
+          <Route exact path="/signup" component={Signup} />
+          <Route exact path="/login" component={Login} />
+          <Route exact path="/profile" component={Profile} />
+          <Route exact path="/challenge" component={Challenge} />
+        </Switch>
+      </Fragment>
     );
   }
 }
